Added email retrieval script
I want a way of returning the contents of the registration table to
save keep logging in so I’ve added a script. It’s in a different area
because I don’t want someone stumbling on it, but it is far from safe
so I should consider removing it the second that we start generating
interest.